Simplify stress tests
#139
