+++ DnDGen.CharacterGen.Tests.Integration.Stress/Randomizers/Abilities/TwoTenSidedDiceAbilitiesRandomizerTests.cs
@@ -33,20 +33,7 @@ namespace DnDGen.CharacterGen.Tests.Integration.Stress.Randomizers.Abilities
             Assert.That(stats.Keys, Contains.Item(AbilityConstants.Intelligence));
             Assert.That(stats.Keys, Contains.Item(AbilityConstants.Strength));
             Assert.That(stats.Keys, Contains.Item(AbilityConstants.Wisdom));
-
-            foreach (var kvp in stats)
-            {
-                Assert.That(kvp.Value.Name, Is.EqualTo(kvp.Key));
-                Assert.That(kvp.Value.Value, Is.InRange(2, 20), kvp.Key);
-            }
-        }
-
-        [Test]
-        public void NonDefaultTwoTenSidedDiceAbilitiesOccur()
-        {
-            var stats = stressor.GenerateOrFail(twoTenSidedDiceAbilitiesRandomizer.Randomize, ss => ss.Values.Any(s => s.Value != 10));
-            var allAbilitiesAreDefault = stats.Values.All(s => s.Value == 10);
-            Assert.That(allAbilitiesAreDefault, Is.False);
+            Assert.That(stats.Values.Select(s => s.Value), Is.All.InRange(2, 20));
         }
     }
 }
